+++ Marketplace/purchase.backend/src/main/java/co/edu/uniandes/csw/miso4204/purchase/logic/ejb/PurchaseLogicService.java
@@ -31,12 +31,12 @@ Source generated by CrudMaker version 1.0.0.qualifier
 package co.edu.uniandes.csw.miso4204.purchase.logic.ejb;
 
 import co.edu.uniandes.csw.miso4204.purchase.logic.dto.PurchasePageDTO;
-import co.edu.uniandes.csw.miso4204.purchase.master.persistence.PurchaseMasterPersistence;
+import co.edu.uniandes.csw.miso4204.purchase.persistence.PurchasePersistence;
 
 public class PurchaseLogicService extends _PurchaseLogicService {
 
 	public PurchaseLogicService() {
-		persistance = new PurchaseMasterPersistence();
+		persistance = new PurchasePersistence();
 	}
 	
         public PurchasePageDTO getPurchasesSearch(Integer page, Integer maxRecords, String ini_date, String end_date, Long id_purchase){
