+++ Marketplace/purchase.backend/src/main/java/co/edu/uniandes/csw/miso4204/purchase/master/logic/ejb/PurchaseMasterLogicService.java
@@ -29,12 +29,14 @@ Source generated by CrudMaker version 1.0.0.qualifier
 */
 
 package co.edu.uniandes.csw.miso4204.purchase.master.logic.ejb;
+import co.edu.uniandes.csw.miso4204.purchase.persistence.PurchasePersistence;
 import co.edu.uniandes.csw.miso4204.purchase.master.persistence.PurchaseMasterPersistence;
 import co.edu.uniandes.csw.miso4204.purchaseitem.persistence.PurchaseItemPersistence;
 import co.edu.uniandes.csw.miso4204.payment.persistence.PaymentPersistence;
 public class PurchaseMasterLogicService extends _PurchaseMasterLogicService {
 
 public PurchaseMasterLogicService() {
+		purchasePersistance = new PurchasePersistence();
 		purchaseMasterPersistance = new PurchaseMasterPersistence();
 		purchaseItemPersistance = new PurchaseItemPersistence();
 		paymentPersistance = new PaymentPersistence();
