+++ Marketplace/purchase.backend/src/main/java/co/edu/uniandes/csw/miso4204/purchase/master/persistence/_PurchaseMasterPersistence.java
@@ -4,31 +4,31 @@
  * Licensed under the MIT, The MIT License (MIT)
  * Copyright (c) 2014 miso4204
 
- Permission is hereby granted, free of charge, to any person obtaining a copy
- of this software and associated documentation files (the "Software"), to deal
- in the Software without restriction, including without limitation the rights
- to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- copies of the Software, and to permit persons to whom the Software is
- furnished to do so, subject to the following conditions:
+Permission is hereby granted, free of charge, to any person obtaining a copy
+of this software and associated documentation files (the "Software"), to deal
+in the Software without restriction, including without limitation the rights
+to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+copies of the Software, and to permit persons to whom the Software is
+furnished to do so, subject to the following conditions:
 
- The above copyright notice and this permission notice shall be included in
- all copies or substantial portions of the Software.
+The above copyright notice and this permission notice shall be included in
+all copies or substantial portions of the Software.
 
- THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- THE SOFTWARE.
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+THE SOFTWARE.
  * ========================================================================
 
 
- Source generated by CrudMaker version 1.0.0.qualifier
+Source generated by CrudMaker version 1.0.0.qualifier
 
- */
+*/
+
 package co.edu.uniandes.csw.miso4204.purchase.master.persistence;
-
 import co.edu.uniandes.csw.miso4204.purchaseitem.logic.dto.PurchaseItemDTO;
 import co.edu.uniandes.csw.miso4204.purchase.master.persistence.entity.PurchasepurchaseItemEntity;
 import co.edu.uniandes.csw.miso4204.purchaseitem.persistence.converter.PurchaseItemConverter;
@@ -36,24 +36,23 @@ import co.edu.uniandes.csw.miso4204.payment.logic.dto.PaymentDTO;
 import co.edu.uniandes.csw.miso4204.purchase.master.persistence.entity.PurchasepaymentEntity;
 import co.edu.uniandes.csw.miso4204.payment.persistence.converter.PaymentConverter;
 import co.edu.uniandes.csw.miso4204.purchase.logic.dto.PurchaseDTO;
-import co.edu.uniandes.csw.miso4204.purchase.logic.dto.PurchasePageDTO;
 import co.edu.uniandes.csw.miso4204.purchase.master.logic.dto.PurchaseMasterDTO;
-import co.edu.uniandes.csw.miso4204.purchase.master.persistence.entity.PurchaseEntity;
-import co.edu.uniandes.csw.miso4204.purchase.persistence.converter.PurchaseConverter;
+import co.edu.uniandes.csw.miso4204.purchase.persistence.PurchasePersistence;
 
 import java.util.ArrayList;
 import java.util.List;
 import javax.persistence.EntityManager;
 import javax.persistence.Query;
 
-public class _PurchaseMasterPersistence {
+public class _PurchaseMasterPersistence{
 
     protected EntityManager entityManager;
+    
+    protected PurchasePersistence purchasePersistence;  
 
-
-    public PurchaseMasterDTO getMasterPurchase(Long purchaseId) {
+    public PurchaseMasterDTO getPurchase(Long purchaseId) {
         PurchaseMasterDTO purchaseMasterDTO = new PurchaseMasterDTO();
-        PurchaseDTO purchase = getPurchase(purchaseId);
+        PurchaseDTO purchase = purchasePersistence.getPurchase(purchaseId);
         purchaseMasterDTO.setPurchaseEntity(purchase);
         purchaseMasterDTO.setListpurchaseItem(getPurchasepurchaseItemEntityList(purchaseId));
         purchaseMasterDTO.setListpayment(getPurchasepaymentEntityList(purchaseId));
@@ -62,125 +61,67 @@ public class _PurchaseMasterPersistence {
 
     public PurchasepurchaseItemEntity createPurchasepurchaseItemEntity(PurchasepurchaseItemEntity entity) {
         entityManager.getTransaction().begin();
-        entityManager.persist(entity);
-        entityManager.getTransaction().commit();
+		entityManager.persist(entity);
+		entityManager.getTransaction().commit();
         return entity;
     }
 
-    public void deletePurchasepurchaseItemEntity(Long purchaseId, Long purchaseItemId) {
-        entityManager.getTransaction().begin();
+   public void deletePurchasepurchaseItemEntity(Long purchaseId, Long purchaseItemId) {
+		entityManager.getTransaction().begin();
         Query q = entityManager.createNamedQuery("PurchasepurchaseItemEntity.deletePurchasepurchaseItemEntity");
         q.setParameter("purchaseId", purchaseId);
         q.setParameter("purchaseItemId", purchaseItemId);
         q.executeUpdate();
-        entityManager.getTransaction().commit();
+		entityManager.getTransaction().commit();
     }
 
-    public List<PurchaseItemDTO> getPurchasepurchaseItemEntityList(Long purchaseId) {
-        entityManager.getTransaction().begin();
-        ArrayList<PurchaseItemDTO> resp = new ArrayList<PurchaseItemDTO>();
+   public List<PurchaseItemDTO> getPurchasepurchaseItemEntityList(Long purchaseId) {
+		entityManager.getTransaction().begin();
+		ArrayList<PurchaseItemDTO> resp = new ArrayList<PurchaseItemDTO>();
         Query q = entityManager.createNamedQuery("PurchasepurchaseItemEntity.getByMasterId");
-        q.setParameter("purchaseId", purchaseId);
-        List<PurchasepurchaseItemEntity> qResult = q.getResultList();
-        for (PurchasepurchaseItemEntity entity : qResult) {
-            if (entity.getPurchaseItemIdEntity() == null) {
+        q.setParameter("purchaseId",purchaseId);
+        List<PurchasepurchaseItemEntity> qResult =  q.getResultList();
+        for (PurchasepurchaseItemEntity entity : qResult) { 
+            if(entity.getPurchaseItemIdEntity()==null){
                 entityManager.refresh(entity);
             }
             resp.add(PurchaseItemConverter.entity2PersistenceDTO(entity.getPurchaseItemIdEntity()));
         }
         entityManager.getTransaction().commit();
         return resp;
-
+      
     }
-
     public PurchasepaymentEntity createPurchasepaymentEntity(PurchasepaymentEntity entity) {
         entityManager.getTransaction().begin();
-        entityManager.persist(entity);
-        entityManager.getTransaction().commit();
+		entityManager.persist(entity);
+		entityManager.getTransaction().commit();
         return entity;
     }
 
-    public void deletePurchasepaymentEntity(Long purchaseId, Long paymentId) {
-        entityManager.getTransaction().begin();
+   public void deletePurchasepaymentEntity(Long purchaseId, Long paymentId) {
+		entityManager.getTransaction().begin();
         Query q = entityManager.createNamedQuery("PurchasepaymentEntity.deletePurchasepaymentEntity");
         q.setParameter("purchaseId", purchaseId);
         q.setParameter("paymentId", paymentId);
         q.executeUpdate();
-        entityManager.getTransaction().commit();
+		entityManager.getTransaction().commit();
     }
 
-    public List<PaymentDTO> getPurchasepaymentEntityList(Long purchaseId) {
-        entityManager.getTransaction().begin();
-        ArrayList<PaymentDTO> resp = new ArrayList<PaymentDTO>();
+   public List<PaymentDTO> getPurchasepaymentEntityList(Long purchaseId) {
+		entityManager.getTransaction().begin();
+		ArrayList<PaymentDTO> resp = new ArrayList<PaymentDTO>();
         Query q = entityManager.createNamedQuery("PurchasepaymentEntity.getByMasterId");
-        q.setParameter("purchaseId", purchaseId);
-        List<PurchasepaymentEntity> qResult = q.getResultList();
-        for (PurchasepaymentEntity entity : qResult) {
-            if (entity.getPaymentIdEntity() == null) {
+        q.setParameter("purchaseId",purchaseId);
+        List<PurchasepaymentEntity> qResult =  q.getResultList();
+        for (PurchasepaymentEntity entity : qResult) { 
+            if(entity.getPaymentIdEntity()==null){
                 entityManager.refresh(entity);
             }
             resp.add(PaymentConverter.entity2PersistenceDTO(entity.getPaymentIdEntity()));
         }
         entityManager.getTransaction().commit();
         return resp;
-
+      
     }
-
-    public PurchaseDTO createPurchase(PurchaseDTO purchase) {
-        PurchaseEntity entity = PurchaseConverter.persistenceDTO2Entity(purchase);
-        entityManager.getTransaction().begin();
-        entityManager.persist(entity);
-        entityManager.getTransaction().commit();
-        return PurchaseConverter.entity2PersistenceDTO(entity);
-    }
-
-    @SuppressWarnings("unchecked")
-    public List<PurchaseDTO> getPurchases() {
-        entityManager.getTransaction().begin();
-        Query q = entityManager.createQuery("select u from PurchaseEntity u");
-        List<PurchaseDTO> result = PurchaseConverter.entity2PersistenceDTOList(q.getResultList());
-        entityManager.getTransaction().commit();
-        return result;
-
-    }
-
-    @SuppressWarnings("unchecked")
-    public PurchasePageDTO getPurchases(Integer page, Integer maxRecords) {
-        entityManager.getTransaction().begin();
-        Query count = entityManager.createQuery("select count(u) from PurchaseEntity u");
-        Long regCount = 0L;
-        regCount = Long.parseLong(count.getSingleResult().toString());
-        Query q = entityManager.createQuery("select u from PurchaseEntity u");
-        if (page != null && maxRecords != null) {
-            q.setFirstResult((page - 1) * maxRecords);
-            q.setMaxResults(maxRecords);
-        }
-        PurchasePageDTO response = new PurchasePageDTO();
-        response.setTotalRecords(regCount);
-        response.setRecords(PurchaseConverter.entity2PersistenceDTOList(q.getResultList()));
-        entityManager.getTransaction().commit();
-        return response;
-    }
-
-    public PurchaseDTO getPurchase(Long id) {
-        entityManager.getTransaction().begin();
-        PurchaseDTO result = PurchaseConverter.entity2PersistenceDTO(entityManager.find(PurchaseEntity.class, id));
-        entityManager.getTransaction().commit();
-        return result;
-    }
-
-    public void deletePurchase(Long id) {
-        entityManager.getTransaction().begin();
-        PurchaseEntity entity = entityManager.find(PurchaseEntity.class, id);
-        entityManager.remove(entity);
-        entityManager.getTransaction().commit();
-    }
-
-    public void updatePurchase(PurchaseDTO detail) {
-        entityManager.getTransaction().begin();
-        PurchaseEntity entity = entityManager.merge(PurchaseConverter.persistenceDTO2Entity(detail));
-        PurchaseConverter.entity2PersistenceDTO(entity);
-        entityManager.getTransaction().commit();
-    }
-
+ 
 }
