+++ Marketplace/purchase.backend/src/main/java/co/edu/uniandes/csw/miso4204/purchase/master/persistence/entity/PurchasepaymentEntity.java
@@ -31,6 +31,7 @@ Source generated by CrudMaker version 1.0.0.qualifier
 package co.edu.uniandes.csw.miso4204.purchase.master.persistence.entity;
 
 import co.edu.uniandes.csw.miso4204.payment.persistence.entity.PaymentEntity;
+import co.edu.uniandes.csw.miso4204.purchase.persistence.entity.PurchaseEntity;
 import java.io.Serializable;
 import javax.persistence.Column;
 import javax.persistence.Entity;
@@ -41,12 +42,8 @@ import javax.persistence.NamedQueries;
 import javax.persistence.NamedQuery;
 import javax.persistence.PrimaryKeyJoinColumn; 
 import org.eclipse.persistence.annotations.JoinFetch;
-import org.eclipse.persistence.annotations.Multitenant;
-import org.eclipse.persistence.annotations.TenantDiscriminatorColumn;
 
 @Entity
-@Multitenant
-@TenantDiscriminatorColumn(name="tenantID")
 @IdClass(PurchasepaymentEntityId.class)
 @NamedQueries({
     @NamedQuery(name = "PurchasepaymentEntity.getByMasterId", query = "SELECT  u FROM PurchasepaymentEntity u WHERE u.purchaseId=:purchaseId"),
