+++ Marketplace/purchase.backend/src/main/java/co/edu/uniandes/csw/miso4204/purchase/master/persistence/entity/PurchasepurchaseItemEntity.java
@@ -31,6 +31,7 @@ Source generated by CrudMaker version 1.0.0.qualifier
 package co.edu.uniandes.csw.miso4204.purchase.master.persistence.entity;
 
 import co.edu.uniandes.csw.miso4204.purchaseitem.persistence.entity.PurchaseItemEntity;
+import co.edu.uniandes.csw.miso4204.purchase.persistence.entity.PurchaseEntity;
 import java.io.Serializable;
 import javax.persistence.Column;
 import javax.persistence.Entity;
@@ -40,14 +41,9 @@ import javax.persistence.ManyToOne;
 import javax.persistence.NamedQueries;
 import javax.persistence.NamedQuery;
 import javax.persistence.PrimaryKeyJoinColumn; 
-import javax.persistence.Transient;
 import org.eclipse.persistence.annotations.JoinFetch;
-import org.eclipse.persistence.annotations.Multitenant;
-import org.eclipse.persistence.annotations.TenantDiscriminatorColumn;
 
 @Entity
-@Multitenant
-@TenantDiscriminatorColumn(name="tenantID")
 @IdClass(PurchasepurchaseItemEntityId.class)
 @NamedQueries({
     @NamedQuery(name = "PurchasepurchaseItemEntity.getByMasterId", query = "SELECT  u FROM PurchasepurchaseItemEntity u WHERE u.purchaseId=:purchaseId"),
@@ -58,16 +54,13 @@ public class PurchasepurchaseItemEntity implements Serializable {
     @Id
     @Column(name = "purchaseId")
     private Long purchaseId;
-    
     @Id
     @Column(name = "purchaseItemId")
     private Long purchaseItemId;
-    
     @ManyToOne
     @PrimaryKeyJoinColumn(name = "purchaseId", referencedColumnName = "id")
     @JoinFetch
-    private PurchaseEntity purchaseIdEntity;
-    
+    private PurchaseItemEntity purchaseIdEntity;
     @ManyToOne
     @PrimaryKeyJoinColumn(name = "purchaseItemId", referencedColumnName = "id")
     @JoinFetch
@@ -97,11 +90,11 @@ public class PurchasepurchaseItemEntity implements Serializable {
         this.purchaseItemId = purchaseItemId;
     }
 
-    public PurchaseEntity getPurchaseIdEntity() {
+    public PurchaseItemEntity getPurchaseIdEntity() {
         return purchaseIdEntity;
     }
 
-    public void setPurchaseIdEntity(PurchaseEntity purchaseIdEntity) {
+    public void setPurchaseIdEntity(PurchaseItemEntity purchaseIdEntity) {
         this.purchaseIdEntity = purchaseIdEntity;
     }
