+++ Marketplace/purchase.backend/src/main/java/co/edu/uniandes/csw/miso4204/purchase/master/persistence/entity/PurchasepurchaseItemEntityId.java
@@ -31,13 +31,11 @@ Source generated by CrudMaker version 1.0.0.qualifier
 package co.edu.uniandes.csw.miso4204.purchase.master.persistence.entity;
 
 import java.io.Serializable;
-import javax.persistence.Embeddable;
 
 /**
  *
  * @author 
  */
-@Embeddable
 public class PurchasepurchaseItemEntityId implements Serializable{
 
     private Long purchaseId;
