--- Marketplace/purchase.backend/src/main/java/co/edu/uniandes/csw/miso4204/purchase/master/persistence/entity/_PurchaseEntity.java
@@ -1,124 +1,0 @@
-/* ========================================================================
- * Copyright 2014 miso4204
- *
- * Licensed under the MIT, The MIT License (MIT)
- * Copyright (c) 2014 miso4204
-
-Permission is hereby granted, free of charge, to any person obtaining a copy
-of this software and associated documentation files (the "Software"), to deal
-in the Software without restriction, including without limitation the rights
-to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-copies of the Software, and to permit persons to whom the Software is
-furnished to do so, subject to the following conditions:
-
-The above copyright notice and this permission notice shall be included in
-all copies or substantial portions of the Software.
-
-THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
-THE SOFTWARE.
- * ========================================================================
-
-
-Source generated by CrudMaker version 1.0.0.qualifier
-
-*/
-
-package co.edu.uniandes.csw.miso4204.purchase.master.persistence.entity;
-
-import java.util.Date;
-
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-import javax.persistence.Temporal;
-import javax.persistence.TemporalType;
-import javax.persistence.MappedSuperclass;
-import javax.persistence.Transient;
-
-@MappedSuperclass
-public abstract class _PurchaseEntity {
-	
-	@Transient
-	private Long componentId;
-	
-	public Long getComponentId() {
-		return componentId;
-	}
-
-	public void setComponentId(Long componentId) {
-		this.componentId = componentId;
-	}
-
-	@Id
-	@GeneratedValue(generator = "Purchase")
-	private Long id;
-	private String name;
-	@Temporal(TemporalType.DATE)
-	private Date purchaseDate;
-	private Double totalValue;
-	private Integer totalItems;
-	private Integer points;
-	private Long buyerId;
-	private Long addressId;
-
-	public Long getId(){
-		return id;
-	}
-	
-	public void setId(Long id){
-		this.id = id;
-	}
-	public String getName(){
-		return name;
-	}
-	
-	public void setName(String name){
-		this.name = name;
-	}
-	public Date getPurchaseDate(){
-		return purchaseDate;
-	}
-	
-	public void setPurchaseDate(Date purchaseDate){
-		this.purchaseDate = purchaseDate;
-	}
-	public Double getTotalValue(){
-		return totalValue;
-	}
-	
-	public void setTotalValue(Double totalValue){
-		this.totalValue = totalValue;
-	}
-	public Integer getTotalItems(){
-		return totalItems;
-	}
-	
-	public void setTotalItems(Integer totalItems){
-		this.totalItems = totalItems;
-	}
-	public Integer getPoints(){
-		return points;
-	}
-	
-	public void setPoints(Integer points){
-		this.points = points;
-	}
-	public Long getBuyerId(){
-		return buyerId;
-	}
-	
-	public void setBuyerId(Long buyerId){
-		this.buyerId = buyerId;
-	}
-	public Long getAddressId(){
-		return addressId;
-	}
-	
-	public void setAddressId(Long addressId){
-		this.addressId = addressId;
-	}
-}
