+++ Marketplace/purchase.backend/src/main/java/co/edu/uniandes/csw/miso4204/purchase/persistence/_PurchasePersistence.java
@@ -1,0 +1,105 @@
+/* ========================================================================
+ * Copyright 2014 miso4204
+ *
+ * Licensed under the MIT, The MIT License (MIT)
+ * Copyright (c) 2014 miso4204
+
+Permission is hereby granted, free of charge, to any person obtaining a copy
+of this software and associated documentation files (the "Software"), to deal
+in the Software without restriction, including without limitation the rights
+to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+copies of the Software, and to permit persons to whom the Software is
+furnished to do so, subject to the following conditions:
+
+The above copyright notice and this permission notice shall be included in
+all copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+THE SOFTWARE.
+ * ========================================================================
+
+
+Source generated by CrudMaker version 1.0.0.qualifier
+
+*/
+
+package co.edu.uniandes.csw.miso4204.purchase.persistence;
+
+import java.util.List;
+
+import javax.persistence.EntityManager;
+import javax.persistence.Query;
+
+import co.edu.uniandes.csw.miso4204.purchase.logic.dto.PurchaseDTO;
+import co.edu.uniandes.csw.miso4204.purchase.logic.dto.PurchasePageDTO;
+import co.edu.uniandes.csw.miso4204.purchase.persistence.converter.PurchaseConverter;
+import co.edu.uniandes.csw.miso4204.purchase.persistence.entity.PurchaseEntity;
+import java.util.Date;
+
+public abstract class _PurchasePersistence{
+
+  	protected EntityManager entityManager;
+	
+	public PurchaseDTO createPurchase(PurchaseDTO purchase) {
+		PurchaseEntity entity=PurchaseConverter.persistenceDTO2Entity(purchase);
+		entityManager.getTransaction().begin();
+		entityManager.persist(entity);
+		entityManager.getTransaction().commit();
+		return PurchaseConverter.entity2PersistenceDTO(entity);
+	}
+	
+	@SuppressWarnings("unchecked")
+	public List<PurchaseDTO> getPurchases() {
+		entityManager.getTransaction().begin();
+		Query q = entityManager.createQuery("select u from PurchaseEntity u");
+		List<PurchaseDTO> result = PurchaseConverter.entity2PersistenceDTOList(q.getResultList());
+		entityManager.getTransaction().commit();
+		return result;
+		
+	}
+
+	@SuppressWarnings("unchecked")
+	public PurchasePageDTO getPurchases(Integer page, Integer maxRecords) {
+		entityManager.getTransaction().begin();
+		Query count = entityManager.createQuery("select count(u) from PurchaseEntity u");
+		Long regCount = 0L;
+		regCount = Long.parseLong(count.getSingleResult().toString());
+		Query q = entityManager.createQuery("select u from PurchaseEntity u");
+		if (page != null && maxRecords != null) {
+		    q.setFirstResult((page-1)*maxRecords);
+		    q.setMaxResults(maxRecords);
+		}
+		PurchasePageDTO response = new PurchasePageDTO();
+		response.setTotalRecords(regCount);
+		response.setRecords(PurchaseConverter.entity2PersistenceDTOList(q.getResultList()));
+		entityManager.getTransaction().commit();
+		return response;
+	}
+        
+	public PurchaseDTO getPurchase(Long id) {
+		entityManager.getTransaction().begin();
+		PurchaseDTO result = PurchaseConverter.entity2PersistenceDTO(entityManager.find(PurchaseEntity.class, id));
+		entityManager.getTransaction().commit();
+		return result;
+	}
+
+	public void deletePurchase(Long id) {
+		entityManager.getTransaction().begin();
+		PurchaseEntity entity=entityManager.find(PurchaseEntity.class, id);
+		entityManager.remove(entity);
+		entityManager.getTransaction().commit();
+	}
+
+	public void updatePurchase(PurchaseDTO detail) {
+		entityManager.getTransaction().begin();
+		PurchaseEntity entity=entityManager.merge(PurchaseConverter.persistenceDTO2Entity(detail));
+		PurchaseConverter.entity2PersistenceDTO(entity);
+		entityManager.getTransaction().commit();
+	}
+
+}
