+++ Marketplace/purchase.services/pom.xml
@@ -81,11 +81,6 @@
             <artifactId>jersey-spring3</artifactId>
             <version>2.13</version>
         </dependency>
-        <dependency>
-            <groupId>${project.groupId}</groupId>
-            <artifactId>security.subsystem</artifactId>
-            <version>1.0-SNAPSHOT</version>
-        </dependency>
     </dependencies>
     <build>
 		<plugins>
