+++ Marketplace/purchaseitem.services/pom.xml
@@ -87,11 +87,8 @@
 			            <version>0.0.1-SNAPSHOT</version>
 			            <type>war</type>
 			        </dependency>
-        <dependency>
-            <groupId>${project.groupId}</groupId>
-            <artifactId>security.subsystem</artifactId>
-            <version>1.0-SNAPSHOT</version>
-        </dependency>
+ 	
+	
     </dependencies>
     <build>
 		<plugins>
